+++ components/molecules/src/GlowLogo/GlowLogo.tsx
@@ -1,5 +1,6 @@
-import Image, { ImageProps } from "next/image";
-import { FC } from "react";
+import type { ImageProps } from "next/image";
+import Image from "next/image";
+import type { FC } from "react";
 
 import styles from "./GlowLogo.module.css";
 
@@ -7,7 +8,7 @@ type LogoGlowProps = Pick<ImageProps, "src">;
 
 export const GlowLogo: FC<LogoGlowProps> = ({ src }) => {
   return (
-    <div className="relative w-full">
+    <div className="relative w-full animate-fade-in-down">
       <Image
         priority
         alt="logo"