+++ components/screens/src/index.ts
@@ -1,3 +1,4 @@
 export { ButtonScreen } from "./ButtonScreen";
 export { LandingDaoboardScreen } from "./LandingDaoboardScreen";
+export { MetafamLandingScreen } from "./MetafamLandingScreen";
 export { LandingMoshScreen } from "./LandingMoshScreen";