+++ configurations/tailwindcss/src/index.js
@@ -14,14 +14,17 @@ module.exports = {
   ],
   purge: [
     join(__dirname, "src/**/*.{js,ts,jsx,tsx}"),
-    join(__dirname, "../../../components/atoms/src/**/*.{js,ts,jsx,tsx}"),
-    join(__dirname, "../../../components/molecules/src/**/*.{js,ts,jsx,tsx}"),
-    join(__dirname, "../../../components/organisms/src/**/*.{js,ts,jsx,tsx}"),
-    join(__dirname, "../../../components/roots/src/**/*.{js,ts,jsx,tsx}"),
-    join(__dirname, "../../../components/screens/src/**/*.{js,ts,jsx,tsx}"),
+    join(__dirname, "../../../apps/**/src/**/*.{js,ts,jsx,tsx}"),
+    join(__dirname, "../../../components/**/src/**/*.{js,ts,jsx,tsx}"),
   ],
   theme: {
     extend: {
+      animation: {
+        "fade-in-down": "fade-in-down 0.5s ease-out",
+        "fade-out-down": "fade-out-down 0.5s ease-out",
+        "fade-in-up": "fade-in-up 0.5s ease-out",
+        "fade-out-up": "fade-out-up 0.5s ease-out",
+      },
       colors: {
         amber: colors.amber,
         blueGray: colors.blueGray,
@@ -39,6 +42,48 @@ module.exports = {
         trueGray: colors.trueGray,
         violet: colors.violet,
         warmGray: colors.warmGray,
+      },
+      keyframes: {
+        "fade-in-down": {
+          "0%": {
+            opacity: "0",
+            transform: "translateY(-10px)",
+          },
+          "100%": {
+            opacity: "1",
+            transform: "translateY(0)",
+          },
+        },
+        "fade-out-down": {
+          from: {
+            opacity: "1",
+            transform: "translateY(0px)",
+          },
+          to: {
+            opacity: "0",
+            transform: "translateY(10px)",
+          },
+        },
+        "fade-in-up": {
+          "0%": {
+            opacity: "0",
+            transform: "translateY(10px)",
+          },
+          "100%": {
+            opacity: "1",
+            transform: "translateY(0)",
+          },
+        },
+        "fade-out-up": {
+          from: {
+            opacity: "1",
+            transform: "translateY(0px)",
+          },
+          to: {
+            opacity: "0",
+            transform: "translateY(10px)",
+          },
+        },
       },
     },
   },