+++ jest.preset.js
@@ -13,6 +13,8 @@ module.exports = {
   moduleNameMapper: {
     "\\.(jpg|ico|jpeg|png|gif|eot|otf|webp|svg|ttf|woff|woff2|mp4|webm|wav|mp3|m4a|aac|oga)$":
       "<rootDir>/../../tests/__mocks__/fileMock.js",
+    "\\.(css|less|scss|sss|styl)$":
+      "<rootDir>/../../node_modules/jest-css-modules",
     "^@sentrei/atoms": "<rootDir>/../../components/atoms/src/index.ts",
     "^@sentrei/molecules": "<rootDir>/../../components/molecules/src/index.ts",
     "^@sentrei/organisms": "<rootDir>/../../components/organisms/src/index.ts",