+++ script.sh
@@ -48,7 +48,12 @@ if [[ $APP ]]; then
       exit 1
   elif [ $GITHUB_ACTIONS ]; then
       echo "✅ - Build can proceed in github actions at $APP - $CHANGED"
-      exit 1
+      yarn run build:$APP
+      if [[ "$GITHUB_EVENT_NAME" == "push" ]]; then
+        cd apps/$APP
+        yarn run serverless
+      fi
+      exit 0
   else
     echo "🌼 - Build not proceeding at $APP - $CHANGED"
     exit 0