+++ configurations/tailwindcss/CHANGELOG.md
@@ -1,3 +1,9 @@
+# @sentrei/tailwindcss [1.3.0](https://github.com/sentrei/sentrei/compare/@sentrei/tailwindcss@1.2.0...@sentrei/tailwindcss@1.3.0) (2021-09-23)
+
+### Features
+
+- ini app mosh lol with json sort ([3bd1255](https://github.com/sentrei/sentrei/commit/3bd12550f6f1a2be250c0497c665e79e9d1ecd88))
+
 # @sentrei/tailwindcss [1.2.0](https://github.com/sentrei/sentrei/compare/@sentrei/tailwindcss@1.1.3...@sentrei/tailwindcss@1.2.0) (2021-09-21)
 
 ### Features